Various script improvements (#12)
* fix: set executable bit for verify_all.sh

* feat: only generate inital phase2 params if needed

If the script was already run and has the correct parameters, don't
recreate them.

* fix: keep .contrib files around

Still delete large phase2 artifacts after successful verification,
bit keep the contributions and their signatures around for later
verification.

* refactor: put base URL into a variable

We are downloading several files. Put the base URL into a variable, so
that it can easily be changed to links to an IPFS gateway.