Verify G1 subgroup and allow verifying contribs independently. (#15)
* Verify G1 subgroup and allow verifying contribs independently.

* fix: don't remove any files

Removing files for saving disk space is an optimization. Don't do that
for now, we care more about correctness.

* fix: make contrib script work correctly

Due to the refactoring `${i}` is now called `${contrib}`.

* fix: don't verify the last parameters again

The final parameter output should have been already be verified
when running `verify_final.sh`.

* fix: proper substraction and input value

`$n-1` == `2-1` for `n='2'`.

The contrib number is called `${contrib}` and not `${n}`. `${n}` is
the total number of contributions.

* fix: use integer comparison

* fix: the `initial_large` value was missing

* fix: improve README instructions

They are not fully correct yet, as it needs to be bumped to this commit.

Co-authored-by: porcuquine <porcuquine@users.noreply.github.com>
Co-authored-by: Volker Mische <volker.mische@gmail.com>