feat: split downloads and verification into two steps (#25)
This is a clear separation of concerns. First get the prerequisites
for the verification, then do the actual verification.

The newly introduced download script is very similar to the
verification script, you can diff between them for easier review
of the code.

There is now a separate script for the initial parameter generation
as well as downloading their prerequisites.