fix: use text mode when checking b2sum checksums
When piping checksum lines to b2sum -c there should be two spaces between the expected digest and filename as opposed to one. This sets the read mode to "text".

Almost all Linux distros, as well as OSX, do not differentiate between text and binary files, and using a single space delimiter will default to text mode. However, it would be better to set this explicitly and to use the line format expected by the standard checksum utilities.